Array iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,13 +1,13 @@
 #include "function_pointers.h"
 /**
- * array_iterator - function that executes a function given as a parameter 
+ * array_iterator - function that executes a function given as a parameter
  * @array: arr
  * @size: size
  * @action: pointer to func
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i = 0;
+	size_t i = 0;
 
 	if (array == NULL || action == NULL)
 	{
